Organize carousel components
